Mostrando datos del carrito de compras para el usuario actual

+++ resources/views/layouts/partials/_header.blade.php
@@ -16,27 +16,37 @@
 					<a class="dropdown-toggle count-info" data-toggle="dropdown" href="#">
 						<i class="fa fa-shopping-cart fa-lg"></i>
 						<span id="cart-count" class="label label-warning">
+						@if (isset($currentCartUser))
+							{{ $currentCartUser->products->count() }}
+						@else
 							0
+						@endif
 						</span>
 					</a>
 					<ul id="products-cart" class="dropdown-menu dropdown-alerts">
-						<li class="li">
-							<div class="row">
-								<div class="col-xs-7 col-sm-7 col-md-7 col-lg-7">
-									<a href="http://tienda-online.local/en/products/show/73"><i class="fa fa-check fa-2x"></i> Rerum</a>
-								</div>
-								<div class="col-xs-1 col-sm-1 col-md-1 col-lg-1">
-									1
-								</div>
-								<div class="col-xs-3 col-sm-3 col-md-3 col-lg-3">
-									<span class="pull-right text-muted small">
-										<a href="http://tienda-online.local/en/cart/api/cart-delete-ajax/73" class="delete-from-cart">
-											<i class="fa fa-minus-circle fa-2x"></i>
+						@if (isset($currentCartUser))
+							@foreach ($currentCartUser->products as $product)
+							<li class="li">
+								<div class="row">
+									<div class="col-xs-7 col-sm-7 col-md-7 col-lg-7">
+										<a href="http://tienda-online.local/en/products/show/73"> 
+											<i class="fa fa-check fa-2x"></i> {{ $product->name }}
 										</a>
-									</span>
+									</div>
+									<div class="col-xs-1 col-sm-1 col-md-1 col-lg-1">
+										{{ $product->pivot->quantity }}
+									</div>
+									<div class="col-xs-3 col-sm-3 col-md-3 col-lg-3">
+										<span class="pull-right text-muted small">
+											<a href="http://tienda-online.local/en/cart/api/cart-delete-ajax/73" class="delete-from-cart">
+												<i class="fa fa-minus-circle fa-2x"></i>
+											</a>
+										</span>
+									</div>
 								</div>
-							</div>
-						</li>
+							</li>
+							@endforeach
+						@endif
 						<li class="divider"></li>
 						<li>
 							<div class="text-center link-block">
